chore: move parser-openapi tests into own file.
I'm about to add some more tests, so figured it was about time they got moved
out of `lib.rs`.
